fix(web-push): move push handlers into sw-src; drop public/ PWA copies (#8)
Upstream landed its own PWA infrastructure while this branch was in flight:
manifest + sw.js are EMITTED by vite.config.ts's emitPwaAssets (standalone
build only, build-fingerprinted for the update prompt) precisely so the embed
island - which copies public/ verbatim - never ships a service worker into a
host app's origin. This branch's public/sw.js + public/manifest.webmanifest
predated that and leaked into the embed build, tripping the
test_embed_build_ships_no_service_worker guard.

- port the push + notificationclick handlers into web/sw-src/sw.js (display
  + routing only; no skipWaiting/claim, so the update-prompt lifecycle is
  untouched)
- delete the public/ sw + manifest (upstream's emitted manifest supersedes)
- drop the eager main.tsx registration (useServiceWorkerUpdate already
  registers /sw.js; enablePushNotifications re-registers idempotently to get
  its registration handle)
