--- web/public/manifest.webmanifest
@@ -1,18 +1,0 @@
-{
-  "name": "Omnigent",
-  "short_name": "Omnigent",
-  "description": "Omnigent — run and manage coding agents.",
-  "start_url": "/",
-  "scope": "/",
-  "display": "standalone",
-  "background_color": "#0a0a0a",
-  "theme_color": "#0a0a0a",
-  "icons": [
-    {
-      "src": "/favicon.svg",
-      "sizes": "any",
-      "type": "image/svg+xml",
-      "purpose": "any maskable"
-    }
-  ]
-}
